+++ R/utils.R
@@ -167,7 +167,7 @@ installed_agg_fun_raw <- function(lib.loc, fields) {
   installed_agg
 }
 
-installed_agg_fun <- memoise::memoise(installed_agg_fun_raw, cache = cachem::cache_mem(max_age = 60 * 60))
+installed_agg_fun <- memoise::memoise(installed_agg_fun_raw, cache = cachem::cache_mem(max_age = 15 * 60))
 
 #' List Available Packages at CRAN-like Repositories
 #' @description available_packages returns a matrix of details corresponding to packages currently available at one or more repositories. The current list of packages is downloaded over the internet (or copied from a local mirror).
@@ -176,13 +176,13 @@ available_packages <- function(repos) {
   tryCatch(available_packages_raw(repos = repos), error = function(e) NA)
 }
 
-available_packages_raw <- memoise::memoise(utils::available.packages, cache = cachem::cache_mem(max_age = 60 * 60))
+available_packages_raw <- memoise::memoise(utils::available.packages, cache = cachem::cache_mem(max_age = 15 * 60))
 
 installed_packages <- function(lib.loc, priority = NULL) {
   installed_packages_raw(lib.loc = lib.loc, priority = priority)
 }
 
-installed_packages_raw <- memoise::memoise(utils::installed.packages, cache = cachem::cache_mem(max_age = 60 * 60))
+installed_packages_raw <- memoise::memoise(utils::installed.packages, cache = cachem::cache_mem(max_age = 15 * 60))
 
 extract_deps <- function(x) {
   splited <- stri_split_fixed(x, ",")
@@ -250,7 +250,7 @@ available_agg_fun_raw <- function(repos, fields) {
   available_agg
 }
 
-available_agg_fun <- memoise::memoise(available_agg_fun_raw, cache = cachem::cache_mem(max_age = 60 * 60))
+available_agg_fun <- memoise::memoise(available_agg_fun_raw, cache = cachem::cache_mem(max_age = 15 * 60))
 
 expand_dependency <- function(x) {
   if (length(x) == 1) {